Protected routes are broken
Before the redirects would occur just based on whether there existed cookies in the header or not. There was no form of authentication with the backend server.

I'm trying to implement a way of checking if the access token or refresh token in the header is valid before redirecting but this is proving to be much harder than I thought it would be :(