Able to give a user Google Authenticator codes
Once the user scans the QR code and enter the 6-digit code, they will be assigned a secret in the database.

Now I want to make it so that if the user has a secret and tries to login, they will need to enter a 6-digit code from Google Authenticator.